refacto: complete restructuration of the project. No code has been changed except for imports. The entire thing is still a WIP
