+++ src/sports_calendar/__main__.py
@@ -1,4 +1,4 @@
-from .cli import app
+from .interfaces.cli import app
 
 if __name__ == "__main__":
     app()
