+++ src/sports_calendar/application/__init__.py
@@ -1,0 +1,2 @@
+import logging
+logger = logging.getLogger(__name__)
