+++ src/sports_calendar/application/selection/__init__.py
@@ -1,4 +1,4 @@
 import logging
 logger = logging.getLogger(__name__)
 
-from .cli import dev_tools
+from .service import SelectionService
