+++ src/sports_calendar/application/workflows/__init__.py
@@ -1,0 +1,5 @@
+import logging
+logger = logging.getLogger(__name__)
+
+from .run_selection import run_selection
+from .clear_calendar import clear_calendar
