+++ src/sports_calendar/core/__init__.py
@@ -2,6 +2,5 @@ import logging
 logger = logging.getLogger(__name__)
 
 from .competition_stages import CompetitionStage
-from .loader import load_yml
-from .setup import Paths, setup_logging
+from ..infra.config import Paths, setup_logging
 from .sports import SportType
