+++ src/sports_calendar/core/selection/__init__.py
@@ -2,7 +2,6 @@ import logging
 logger = logging.getLogger(__name__)
 
 from .models import Selection, SelectionItem
-from .service import SelectionService
 from .filters import (
     SelectionFilter, FilterType, Rule, EntitySelectionRule,
     FilterFields, EmptyFilterFields, MinRankingFilterFields,
