+++ src/sports_calendar/infra/__init__.py
@@ -1,0 +1,2 @@
+import logging
+logger = logging.getLogger(__name__)
