+++ src/sports_calendar/infra/config/paths.py
