+++ src/sports_calendar/infra/engine/executors/competitions.py
