+++ src/sports_calendar/infra/engine/executors/empty.py
