+++ src/sports_calendar/infra/engine/executors/min_ranking.py
