+++ src/sports_calendar/infra/engine/executors/sessions.py
