+++ src/sports_calendar/infra/engine/executors/teams.py
