+++ src/sports_calendar/infra/google_calendar/config/__init__.py
@@ -1,5 +1,3 @@
-import logging
-logger = logging.getLogger(__name__)
-
+from .. import logger
 from .credentials import Credentials
 from .secrets import Secrets
