+++ src/sports_calendar/infra/google_calendar/manager.py
