+++ src/sports_calendar/infra/setup/__init__.py
@@ -1,4 +1,4 @@
 import logging
 logger = logging.getLogger(__name__)
 
-from .main import init
+from .initialize import init
