+++ src/sports_calendar/infra/setup/initialize.py
