+++ src/sports_calendar/infra/storage/__init__.py
@@ -1,0 +1,5 @@
+import logging
+logger = logging.getLogger(__name__)
+
+from .loader import load_yml
+from .selection import SelectionStorage
