+++ src/sports_calendar/infra/storage/loader.py
@@ -3,7 +3,7 @@ from datetime import datetime, timedelta, timezone
 
 import yaml
 
-from . import logger
+from ...core import logger
 
 
 def date_offset_constructor(loader, node):
