+++ src/sports_calendar/interfaces/cli/__init__.py
@@ -1,4 +1,4 @@
 import logging
 logger = logging.getLogger(__name__)
 
-from .cli import sync_calendar, clear_cal
+from .main import app
